handle edge case for strike utils if <= 1 required strikes and <= 0 required min not working days

+++ src/utils/strike_utils.py
@@ -22,19 +22,27 @@ class StrikeUtils:
         Returns:
             bool: True if limit has been reached, false otherwise
         """
+        # If max. 1 required strike and min not working days, criteria is always reached
+        if CONFIG["jobs"][strike_type.value]["required_strikes"] <= 1 and CONFIG["jobs"][strike_type.value]["min_not_working_days"] <= 0:
+            return True
+
         with DbManager() as db:
             row = db.execute_fetchone(query=f"SELECT * FROM {strike_type.value}_strikes WHERE hash = ?", params=(torrent_hash,))
+            # If not existing record, create one
             if not row:
                 logger.trace(f"Torrent with hash {torrent_hash} is not in db and will be created")
                 db.execute(query=f"INSERT INTO {strike_type.value}_strikes (hash, strikes, first_strike) VALUES (?, 1, ?)", params=(torrent_hash, datetime.now()))
                 return False
+            # If existing record, check strikes and days not working
             else:
                 strikes = row["strikes"] + 1
                 strike_days: timedelta = datetime.now() - datetime.strptime(row["first_strike"], "%Y-%m-%d %H:%M:%S.%f")
+                # If required strikes and min not working days are reached, delete torrent and return true
                 if strikes >= CONFIG["jobs"][strike_type.value]["required_strikes"] and strike_days >= timedelta(days=CONFIG["jobs"][strike_type.value]["min_not_working_days"]):
                     logger.trace(f"Torrent with hash {torrent_hash} has reached {strikes} strikes in {strike_days} days - entry is being deleted")
                     db.execute(query=f"DELETE FROM {strike_type.value}_strikes WHERE hash = ?", params=(torrent_hash,))
                     return True
+                # Else, increment strikes
                 logger.trace(f"Torrent with hash {torrent_hash} doesn't meet criteria and will have strikes increased (strikes ({strikes}/{CONFIG["jobs"][strike_type.value]["required_strikes"]}) days ({strike_days}/{CONFIG["jobs"][strike_type.value]["min_not_working_days"]}))")
                 db.execute(query=f"UPDATE {strike_type.value}_strikes SET strikes = strikes + 1 WHERE hash = ?", params=(torrent_hash,))
         return False
